Move WebSocketSessionManager to transport level
