+++ connector-framework/pom.xml
@@ -43,10 +43,6 @@
             <groupId>org.wso2.carbon.messaging</groupId>
             <artifactId>org.wso2.carbon.messaging</artifactId>
         </dependency>
-        <dependency>
-            <groupId>javax.websocket</groupId>
-            <artifactId>javax.websocket-api</artifactId>
-        </dependency>
     </dependencies>
 
     <properties>
@@ -59,7 +55,6 @@
             org.osgi.util.tracker;version="${osgi.service.tracker.import.version.range}",
             org.slf4j.*;version="${slf4j.logging.package.import.version.range}",
             org.wso2.carbon.messaging.*;version="${carbon.messaging.package.import.version.range}",
-            javax.websocket.*;version="${javax.websocket.version.range}"
         </import.package>
     </properties>
