+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/listener/WebSocketSourceHandler.java
@@ -102,6 +102,7 @@ public class WebSocketSourceHandler extends SourceHandler {
             String reasonText = closeWebSocketFrame.reasonText();
             int statusCode = closeWebSocketFrame.statusCode();
             ctx.channel().close();
+            WebSocketSessionManager.getInstance().removeSession(uri, channelId);
             cMsg = new StatusCarbonMessage(org.wso2.carbon.messaging.Constants.STATUS_CLOSE, statusCode, reasonText);
 
         } else if (msg instanceof PongWebSocketFrame) {
@@ -154,6 +155,7 @@ public class WebSocketSourceHandler extends SourceHandler {
         cMsg.setProperty(Constants.CHANNEL_ID, channelId);
         cMsg.setProperty(Constants.CONNECTION, Constants.UPGRADE);
         cMsg.setProperty(Constants.UPGRADE, Constants.WEBSOCKET_UPGRADE);
+        WebSocketSessionManager.getInstance().add(uri, session);
         publishToMessageProcessor(cMsg);
     }
 
@@ -181,5 +183,7 @@ public class WebSocketSourceHandler extends SourceHandler {
         cMsg.setProperty(Constants.REMOTE_PORT, ((InetSocketAddress) ctx.channel().remoteAddress()).getPort());
         cMsg.setProperty(Constants.CHANNEL_ID, channelId);
         cMsg.setProperty(Constants.PROTOCOL, Constants.WEBSOCKET_PROTOCOL);
+        Session session = WebSocketSessionManager.getInstance().getSession(uri, channelId);
+        cMsg.setProperty(Constants.WEBSOCKET_SESSION, session);
     }
 }
